# openapi.yml
openapi: "3.0.0"
info:
  title: User Balance API
  version: "1.0.0"
components:
  schemas:
    transaction:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 505
        user_id:
          type: integer
          example: 5
        amount:
          type: integer
          example: 1000
        type_op:
          description: Тип операции (зачисление или списание средств)
          type: integer
          minimum: 1
          maximum: 2
          example: 1
        transfer_id:
          type: integer
          example: 15
        message:
          type: string
          example: Перевод заработной платы
        created_at:
          type: string
          readOnly: true
          example: 01.01.2022 21:57:05
      required:
        - id
        - user_id
        - amount
        - type_op
        - transfer_id
        - created_at
paths:
  /users/{id}/balance:
    description: Маршрут для работы с балансом пользователей
    get:
      description: Returns balance by user ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
        - in: query
          name: currency
          schema:
            type: string
            example: USD
            enum:
              - USD
      responses:
        "200":
          description: OK Response
          content:
            application/json:
              schema:
                type: object
                items:
                  properties:
                    user_id:
                      type: integer
                      example: 1
                    amount:
                      type: integer
                      example: 1000

  /users/{id}/transactions:
    description: Маршрут для работы с транзакциями пользователей
    get:
      description: Returns transactions by user ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: OK Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/transaction'
  /transactions:
    description: Маршрут для работы с переводами средств
    post:
      description: Creates/edits a transaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transaction'
      responses:
        "200":
          description: OK Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction'